+++ src/components/Dashboard/index.ts
@@ -3,3 +3,4 @@ export * from "./DashboardUserStats";
 export * from "./DashboardFilterGroup";
 export * from "./DashboardNextToJumpBanner";
 export * from "./DashboardBannerRow";
+export * from "./DashboardTableRace";
