+++ src/components/Tables/DashboardTable.tsx
@@ -4,12 +4,11 @@ import { TableData, TableHeader, TableRow } from "../../types/table";
 import { Meet } from "../../types/meets";
 import utils from "../../utils";
 import Skeleton from "react-loading-skeleton";
-import { Link } from "react-router-dom";
-import classnames from "classnames";
 import dayjs from "dayjs";
 import relativeTime from "dayjs/plugin/relativeTime";
 import utc from "dayjs/plugin/utc";
 import advancedFormat from "dayjs/plugin/advancedFormat";
+import { DashboardTableRace } from "../Dashboard";
 
 dayjs.extend(relativeTime);
 dayjs.extend(utc);
@@ -33,55 +32,10 @@ export const DashboardTable: React.FC<Props> = ({ meets }) => {
       title: `${meet.name} (${meet.location})`,
       classNames: "bg-gray-200"
     },
-    ...meet.races.map(race => {
-      const isAfterClosingTime = dayjs().isAfter(dayjs(race.close));
-      return {
-        title: (
-          <Link
-            className={classnames({
-              "!cursor-default":
-                race.status === "Interim" ||
-                race.status === "Abandoned" ||
-                race.status === "Closed" ||
-                (race.status === "Normal" && isAfterClosingTime)
-            })}
-            to={
-              race.status === "Normal" && !isAfterClosingTime
-                ? `/races/${meet.id}/${race.number}`
-                : race.status === "Paying"
-                ? `/results/${utils.markets.getPropositionIdFromRaceMeet(
-                    race,
-                    meet
-                  )}`
-                : // race status in any other condition other than normal or paying
-                  ""
-            }
-          >
-            <div
-              className={classnames("px-3 py-4 whitespace-nowrap text-sm", {
-                "bg-gray-400 hover:bg-gray-500": race.status === "Paying",
-                "bg-black text-white": race.status === "Abandoned",
-                "bg-emerald-400": race.status === "Interim",
-                "hover:bg-gray-200": race.status === "Normal"
-              })}
-            >
-              <p>R{race.number}</p>
-              {dayjs.utc(race.start).local().format("H:mm")}
-              <p>
-                {race.status == "Paying"
-                  ? race.results?.join(" ")
-                  : race.status == "Abandoned"
-                  ? "ABND"
-                  : isAfterClosingTime
-                  ? "CLSD"
-                  : dayjs(race.close).fromNow(true)}
-              </p>
-            </div>
-          </Link>
-        ),
-        classNames: "!p-0"
-      };
-    }),
+    ...meet.races.map(race => ({
+      title: <DashboardTableRace meet={meet} race={race} />,
+      classNames: "!p-0"
+    })),
     ...Array.from({ length: length - meet.races.length }, () => ({
       title: ""
     }))
